Move kafka stuff to infrastructure
