+++ src/main/kotlin/no/nav/syfo/infrastructure/kafka/PersonFlagget84AivenProducer.kt
@@ -1,8 +1,6 @@
-package no.nav.syfo.pengestopp.kafka
+package no.nav.syfo.infrastructure.kafka
 
 import no.nav.syfo.application.Environment
-import no.nav.syfo.application.kafka.JacksonKafkaSerializer
-import no.nav.syfo.application.kafka.commonKafkaAivenConfig
 import no.nav.syfo.pengestopp.StatusEndring
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.apache.kafka.clients.producer.ProducerConfig
