+++ src/main/kotlin/no/nav/syfo/infrastructure/kafka/identhendelse/KafkaIdenthendelseConsumerConfig.kt
@@ -1,9 +1,9 @@
-package no.nav.syfo.identhendelse.kafka
+package no.nav.syfo.infrastructure.kafka.identhendelse
 
 import io.confluent.kafka.serializers.KafkaAvroDeserializer
 import io.confluent.kafka.serializers.KafkaAvroDeserializerConfig
-import no.nav.syfo.application.kafka.KafkaEnvironment
-import no.nav.syfo.application.kafka.commonKafkaAivenConsumerConfig
+import no.nav.syfo.infrastructure.kafka.KafkaEnvironment
+import no.nav.syfo.infrastructure.kafka.commonKafkaAivenConsumerConfig
 import org.apache.kafka.clients.consumer.ConsumerConfig
 import java.util.*
 
@@ -12,7 +12,6 @@ fun kafkaIdenthendelseConsumerConfig(
 ): Properties {
     return Properties().apply {
         putAll(commonKafkaAivenConsumerConfig(kafkaEnvironment))
-        this[ConsumerConfig.GROUP_ID_CONFIG] = "ispengestopp-v1"
         this[ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG] = KafkaAvroDeserializer::class.java.canonicalName
         this[ConsumerConfig.MAX_POLL_RECORDS_CONFIG] = "1"
