+++ src/test/kotlin/no/nav/syfo/testutils/generator/KafkaIdenthendelseDTOGenerator.kt
@@ -1,8 +1,8 @@
 package no.nav.syfo.testutils.generator
 
-import no.nav.syfo.identhendelse.kafka.IdentType
-import no.nav.syfo.identhendelse.kafka.Identifikator
-import no.nav.syfo.identhendelse.kafka.KafkaIdenthendelseDTO
+import no.nav.syfo.infrastructure.kafka.identhendelse.IdentType
+import no.nav.syfo.infrastructure.kafka.identhendelse.Identifikator
+import no.nav.syfo.infrastructure.kafka.identhendelse.KafkaIdenthendelseDTO
 import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.testutils.UserConstants
